backend/requirements.txt
Django==5.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
PyYAML==6.0.1
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.7
defusedxml==0.7.1
django-allauth==65.2.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
flyql_generators @ git+https://github.com/iamtelescope/flyql-generators.git#subdirectory=python
gunicorn==23.0.0
idna==3.7
jsonschema==4.23.0
oauthlib==3.2.2
pycparser==2.22
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
whitenoise==6.6.0
